+++ modules/i18n/src/models/namespaces/applications-ns.ts
@@ -1635,6 +1635,7 @@ export interface ApplicationsNS {
                             valueDescriptions: {
                                 default: string;
                                 jwt: string;
+                                opaque: string;
                             };
                         };
                         revokeToken: {